requirements.txt
PyQt5~=5.15.4
Pillow~=8.3.2
setuptools~=57.4.0